+++ internal/component/mimir/alerts/kubernetes/debug.go
@@ -1,0 +1,6 @@
+package alerts
+
+// TODO: Implement a useful DebugInfo
+func (c *Component) DebugInfo() interface{} {
+	return nil
+}